backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
agora-token-builder==1.0.0
python-dotenv==1.0.1